Remove redundant component
