--- src/web/src/ui-components/Loading.tsx
@@ -1,20 +1,0 @@
-import CircularProgress from "@mui/material/CircularProgress";
-import LinearProgress from "@mui/material/LinearProgress";
-import Typography from "@mui/material/Typography";
-
-export const LoadingAuth = () => (<Loading />);
-export const Loading = (props: {
-  linearProgress?: boolean;
-  label?: string;
-}) => {
-  return (
-    <>
-      <Typography variant="caption" gutterBottom sx={{ display: 'block' }}>{props.label}</Typography>
-      {props.linearProgress ? (
-        <LinearProgress />
-      ) : (
-        <CircularProgress sx={{ padding: 5 }} />
-      )}
-    </>
-  );
-};
